refactor: API creator
